DMResourceHub/requirements.txt
firebase-admin==6.2.0
cloudinary==1.34.0
Pillow==10.0.0
python-dotenv==1.0.0
discord.py==2.3.1
requests==2.31.0
PyPDF2==3.0.1